Revamp of the heart directories structure. Tree rules, current directories are removed. This is known by the master resolve file of a tree.
Tree name directories are removed. Services are not dependent anymore of a tree.
Tree name directories are replaced by a unique services directory containing all services of the system with it resolve file.

Tree resolve file handle now the list of the service name to start with a tree.

This change give the services switch between trees easily without IO access by the simplication
of the research, access of the resolve file and script of a service.