remove -I -f option to enable.
This option come from previous behavior. It not the role of enable process
to deal with parse option.
Also, this change allow user to quickly switch between trees without the need
to force a parse process.
The necessary field at resolve file of the service is now handled by enable
process.