Add the 66-execute program. This program is executed via the run script in the service's live directory. It sits between s6-supervise and the user's executable script, allowing for the preparation of the user's script execution. This introduces a dual capability:
    Define certain preparations through frontend keys, such as setting ownership, moving to a working directory, limiting resources used by the service, etc.
    Since 66-execute runs on each signal received and transmitted by s6-supervise, it is now possible to send messages to an event management daemon.

By default, this program performs previously defined actions such as logging redirections and managing environment variables (previously handled by excel-envfile), creation of the log destination directory, and so on.So, the user's file now contains only the exact copy of the @execute field.
Also, the name of the script for an start or
stop process are the same. This was a old behabvior
coming from s6-rc  manager.